Fix apoc to work with long degrees

+++ core/src/main/java/apoc/stats/DegreeDistribution.java
@@ -60,7 +60,7 @@ public class DegreeDistribution {
         private transient AtomicHistogram histogram;
 
         public void computeDegree(NodeCursor nodeCursor) {
-            int degree = DegreeUtil.degree(nodeCursor, type, direction);
+            long degree = DegreeUtil.degree(nodeCursor, type, direction);
             record(degree);
         }
